+++ ChangeLog.md
@@ -1,6 +1,7 @@
 ### Version v2.3.3
   - CAdminAuthorization: Add pkexec in linux
-
+  - Update: Modify install AppImage
+  
 ### Version v2.3.2
   - Modify install root dir
   - Log: modify log file name